Add: I Added deep links for shareable contract URLs
