+++ src/app/contract/[id]/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { ContractExplorer } from "@/components/contract-explorer";
+
+interface Props {
+  params: {
+    id: string;
+  };
+}
+
+export default function ContractPage({ params }: Props) {
+  return <ContractExplorer initialContractId={params.id} />;
+}
